+++ src/themes/five/src/components/BusinessProductsListing/styles.js
@@ -3,12 +3,6 @@ import styled, { css } from 'styled-components'
 export const ProductsContainer = styled.div`
   width: 95%;
   margin: 20px auto 0px;
-  > svg {
-    color: ${props => props.theme.colors.headingColor};
-    font-size: 24px;
-    cursor: pointer;
-    margin-bottom: 5px;
-  }
 `
 
 export const ProductLoading = styled.div`
@@ -134,4 +128,22 @@ export const EmptyBtnWrapper = styled.div`
     padding-left: 25px;
     padding-right: 25px;
   }
+`
+
+export const HeaderContent = styled.div`
+  display: flex;
+  align-items: center;
+  margin-bottom: 10px;
+  .back-arrow {
+    font-size: 24px;
+    cursor: pointer;
+  }
+`
+export const OrderContextUIWrapper = styled.div`
+  flex: 1;
+  margin-left: 15px;
+  ${props => props.theme.rtl && css`
+    margin-left: 0px;
+    margin-right: 15px;
+  `}
 `
